pyproject.toml
[tool.poetry]
name = "Eric"
version = "0.1.0"
description = ""
authors = ["nullqwertyuiop <nullqwertyuio@outlook.com>"]
license = "AGPL-3.0 license"

[tool.poetry.dependencies]
python = "^3.10"
graia-ariadne = "^0.10.0"
graia-saya = "^0.0.17"
graia-scheduler = "^0.0.10"
graiax-playwright = "^0.2.1"
typing-extensions = "^4.4.0"
SQLAlchemy = "^1.4.44"
aiomysql = "^0.1.1"
aiosqlite = "^0.17.0"
aiohttp = "^3.8.3"
kayaku = "^0.5.0"
fastapi = "^0.88.0"
uvicorn = "^0.20.0"
Pillow = "^9.3.0"
numpy = "^1.23.4"
graiax-fastapi = "^0.2.1"
GitPython = "^3.1.29"
psutil = "^5.9.4"
aiofiles = "^22.1.0"
prompt-toolkit = "^3.0.36"
lxml = "^4.9.1"
graiax-text2img-playwright = "^0.2.0"

[tool.poetry.group.dev.dependencies]
black = "^22.10.0"
pre-commit = "^2.20.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"
